+++ lib/url/shadowsocks.dart
@@ -47,6 +47,13 @@ class ShadowSocksURL extends V2RayURL {
         spiderX: null,
       );
     }
+
+    // Update the null outboundTag in routing rules with server's remark
+    for (var rule in routing['rules']) {
+      if (rule['outboundTag'] == null) {
+        rule['outboundTag'] = remark;
+      }
+    }
   }
 
   @override
@@ -66,38 +73,41 @@ class ShadowSocksURL extends V2RayURL {
 
   @override
   Map<String, dynamic> get outbound1 => {
-        "tag": "proxy",
+        "tag": remark,
         "protocol": "shadowsocks",
         "settings": {
-          "vnext": null,
           "servers": [
             {
               "address": address,
-              "method": method,
-              "ota": false,
-              "password": password,
               "port": port,
-              "level": level,
-              "email": null,
-              "flow": null,
-              "ivCheck": null,
-              "users": null
+              "method": method,
+              "password": password,
+              "uot": true,
+              "level": 0
             }
-          ],
-          "response": null,
-          "network": null,
-          "address": null,
-          "port": null,
-          "domainStrategy": null,
-          "redirect": null,
-          "userLevel": null,
-          "inboundTag": null,
-          "secretKey": null,
-          "peers": null
+          ]
         },
         "streamSettings": streamSetting,
-        "proxySettings": null,
-        "sendThrough": null,
         "mux": {"enabled": false, "concurrency": 8}
       };
+
+  @override
+  Map<String, dynamic> get fullConfiguration {
+    // Ensure routing rules use the correct outboundTag
+    for (var rule in routing['rules']) {
+      if (rule['outboundTag'] == null) {
+        rule['outboundTag'] = remark;
+      }
+    }
+
+    // Add the catch-all rule
+    routing['rules'].add({
+      "type": "field",
+      "port": "0-65535",
+      "outboundTag": remark,
+      "enabled": true
+    });
+
+    return super.fullConfiguration;
+  }
 }