+++ lib/url/vless.dart
@@ -47,7 +47,7 @@ class VlessURL extends V2RayURL {
 
   @override
   Map<String, dynamic> get outbound1 => {
-        "tag": "proxy",
+        "tag": remark,
         "protocol": "vless",
         "settings": {
           "vnext": [
@@ -84,6 +84,26 @@ class VlessURL extends V2RayURL {
         "mux": {
           "enabled": false,
           "concurrency": 8,
-        },
+        }
       };
+
+  @override
+  Map<String, dynamic> get fullConfiguration {
+    // Ensure routing rules use the correct outboundTag
+    for (var rule in routing['rules']) {
+      if (rule['outboundTag'] == null) {
+        rule['outboundTag'] = remark;
+      }
+    }
+
+    // Add the catch-all rule
+    routing['rules'].add({
+      "type": "field",
+      "port": "0-65535",
+      "outboundTag": remark,
+      "enabled": true
+    });
+
+    return super.fullConfiguration;
+  }
 }