+++ lib/url/vmess.dart
@@ -54,7 +54,7 @@ class VmessURL extends V2RayURL {
 
   @override
   Map<String, dynamic> get outbound1 => {
-        "tag": "proxy",
+        "tag": remark,
         "protocol": "vmess",
         "settings": {
           "vnext": [
@@ -95,4 +95,24 @@ class VmessURL extends V2RayURL {
           "concurrency": 8,
         }
       };
+
+  @override
+  Map<String, dynamic> get fullConfiguration {
+    // Ensure routing rules use the correct outboundTag
+    for (var rule in routing['rules']) {
+      if (rule['outboundTag'] == null) {
+        rule['outboundTag'] = remark;
+      }
+    }
+
+    // Add the catch-all rule
+    routing['rules'].add({
+      "type": "field",
+      "port": "0-65535",
+      "outboundTag": remark,
+      "enabled": true
+    });
+
+    return super.fullConfiguration;
+  }
 }